Start to create the GUI
There are some new files and some modified files in the project directory to develop the graphical user interface.
Before this commit, there was just a python file to run the game in the terminal; But now we started to create
a perfect application(including a beautiful GUI and the best features).
Some files added and changed for this purpose:
    - Modified `app.py` to prepare the application to run.
    - Added `gui.py` to write the codes of the UI.
    - Added `options.py` to prepare some new features for the application.
    - Modified `README.md` based on the newest changes.
    - Added `terminal_game.py` to run the simple version of the game in terminal.